resolving warnings & moving to float standard

+++ include/Window.hpp
@@ -19,7 +19,7 @@
 
 class RaytraceWindow {
  public:
-  RaytraceWindow(const size_t screen_width, const size_t screen_height,
+  RaytraceWindow(const int screen_width, const int screen_height,
                  const char* title);
 
   ~RaytraceWindow();
@@ -31,23 +31,23 @@ class RaytraceWindow {
   int current_renders = 0;
   int max_renders = 10;
   uint target_fps = 30;
-  size_t start_index = 0;  // The starting index for the next selection (when
+  int start_index = 0;  // The starting index for the next selection (when
                            // selecting random colors)
 
-  const size_t screen_width;   // The dimensions of the window
-  const size_t screen_height;  // The dimensions of the window
+  const int screen_width;      // The dimensions of the window
+  const int screen_height;     // The dimensions of the window
   ObjectsList* world;          // The world that we are going to draw
   camera cam;                  // The camera object
   std::vector<Color> pixels;   // The array of pixels that we are going to draw
   Texture2D texture;  // The texture that we are going to draw (DrawPixel method
                       // is not efficient for large images)
 
-  std::vector<size_t> shuffled_index_array;  // An array of shuffled indices
+  std::vector<int> shuffled_index_array;  // An array of shuffled indices
 
   void draw_pixels();
   void reset_pixels();
-  double get_ray_random_duration();
-  void shuffle_indices(std::vector<size_t>& indices, size_t start_index);
+  float get_ray_random_duration();
+  void shuffle_indices(std::vector<int>& indices, int start_index);
 };
 
 #endif  // WINDOW_HPP
